LittleReport
Here's a start for reporting to really collect data.  This isn't
intended to be the final draft.  I'd like more feedback or additions to
collecting useful battle data and any recommendations on formatting.  As
of now, it's easy to quickly paste the results in excel but I'd like to
re-tool it to allow a cleaner cut and paste for futher analysis of
everyone's report data.